usdc-hackathon/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
web3==6.18.0
solana-py==0.34.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
prometheus-client==0.19.0
grafana-api==1.0.3
